Rename P2P peer directory PeerList class to BaseList
